Changed IDbTable.delete() return type to Int

+++ app/src/main/java/com/itg3/grp1/mobdevproject/DbTables/IDbTable.kt
@@ -13,5 +13,5 @@ abstract class IDbTable<T: IModel>(newDbHandler: SQLiteOpenHelper)
     abstract fun getAll() : List<T>
     abstract fun add(instance: T) : Long
     abstract fun update(instance: T) : Int
-    abstract fun delete(instance: T) : Long
+    abstract fun delete(instance: T) : Int
 }
